HADOOP-16068. Addendum: ABFS Authentication and Delegation Token plugins to optionally be bound to specific URI of the store.
In CDH6.3 we have refreshed the BouncyCastle dependency version. This patch was included in cdpd with the bcprov-16 used
in the project as a dependency, but after the version increase the link to the BouncyCastle dependency causes a Maven
error here. I am changing the used version as well for the hadoop-azure, as it is in upstream as well in the orig patch.

Change-Id: Ib8e01d3af21f6db3195ab27fb31600d0073fe102