CDPD-2855. HADOOP-16430. S3AFilesystem.delete to incrementally update s3guard with deletions
Contributed by Steve Loughran.

This overlaps the scanning for directory entries with batched calls to S3 DELETE and updates of the S3Guard tables.
It also uses S3Guard to list the files to delete, so find newly created files even when S3 listings are not use consistent.

For path which the client considers S3Guard to be authoritative, we also do a recursive
LIST of the store and delete files; this is to find unindexed files and so guarantee
that the delete(path, true) call really does delete everything underneath.

Contains HADOOP-16554. mvn javadoc:javadoc fails in hadoop-aws which fixes the one-line regression
in javadocs which this patch added.

Change-Id: Ice2f6e940c506e0b3a78fa534a99721b1698708e