CDPD-4271. HADOOP-16579. Upgrade to Apache Curator 4.2.0 and ZooKeeper 3.5.5
- Add ZooKeeper and Curator, excluding dependencies already present
- Add a static initializer for the unit tests using ZooKeeper to enable
the four-letter-words diagnostic telnet commands. (this is an interface
that become disabled by default, so to keep the ZooKeeper 3.4.x behavior
we enabled it for the tests)
- Also fix ZKFailoverController to look for relevant fail-over ActiveAttempt
records. The new ZooKeeper seems to respond quicker during the fail-over
tests than the ZooKeeper, so we made sure to catch all the relevant records
by adding a new parameter to ZKFailoverontroller.waitForActiveAttempt().
- On the new curator some of the API calls we use in Hadoop become
deprecated, I updated that part in hadoop to not have compile time warnings
- I changed the way how some tests are starting ZooKeeper server (the old
server could be started with port 0 and bound to a random open port, but
this feature is not working in ZooKeeper 3.5.5, so I choose a random open
port before starting the ZooKeeper server in the Hadoop code)

Change-Id: I65b3572aa52425e64a5895caf0907bd2b8ea633a