CDPD-30243. HADOOP-16920. ABFS: Make list page size configurable.
Contributed by Bilahari T H.

The page limit is set in "fs.azure.list.max.results"; default value is 500.

There's currently a limit of 5000 in the store -there are no range checks
in the client code so that limit can be changed on the server without
any need to update the abfs connector.

Change-Id: I5cb6ead7aa6c952f15e774850fe6f5dcb1ed3d6e