CDPD-25014. HADOOP-17132. ABFS: Fix Rename and Delete Idempotency check trigger
- Contributed by Sneha Vijayarajan

Conflicts:
	hadoop-tools/hadoop-azure/src/main/java/org/apache/hadoop/fs/azurebfs/services/AbfsClient.java

Backport notes:
In apache hadoop this patch came in before HADOOP-17194 and the AbfsClientContext;
here it is coming in after. Some changes were needed so that the mockito tests
in TestAbfsClient set up the context properly. This was done by pulling in
a helper method from HADOOP-17215 to simplify setting final fields,
and adding the context to the set of fields to update.
This change is *not* in trunk/branch-3.3, which is something
to look at.

Change-Id: Idcd5991dbd66b9d9ad915a0de8163bd148f74cc0