CDPD-18624: Introduce FIPS-mode switch and use it in ProxyCA
The FIPS-mode is determined by the value of JVM property: com.safelogic.cryptocomply.fips.approved_only.
If it is set to "true", FIPS-mode will be enabled.
Enabling FIPS-mode means enforcing that the Safelogic Cryptocomply (CCJ) provider is used
as the default provider for all crypto objects.
There are additional DEBUG / TRACE logs added to have a clear picture of the used providers and crypto objects.

Tested the changes introduced by this commit manually on my FIPS cluster:
- Confirmed that the JVM property enables FIPS-mode.
- Confirmed that all providers for crypto classes are correctly set to CCJ.

Note: This commit is downstream only.

Change-Id: Ic262e89380ae87d7bd9d82ea427af6100183df2e