CDPD-21445 Zeppelin build failure in 7.1.6
Temporarily the dependency jsonschema2pojo was removed. It's unclear why
a different version of com.fasterxml.jackson.dataformat:jackson-dataformat-yaml
appears on the Zeppelin side during the dependency convergence check.

Only 2.10.5 is supposed to be visible. This is a temporary workaround and ideally
we expect things to return to normal as soon as we figure out why this happens.

Change-Id: Iaa5d8ad8fec3767cdc16aff34572f1a84dd66e40