CDPD-26441. HADOOP-17705. S3A to add Config to set AWS region (#3020)
The option `fs.s3a.endpoint.region` can be used
to explicitly set the AWS region of a bucket.

This is needed when using AWS Private Link, as
the region cannot be automatically determined.

Contributed by Mehakmeet Singh

*Important*
This has proven to be a fairly brittle change. Consult the CDPD-26441 JIRA for details.
Test environments with the AWS CLI or run in EC2 do not pick up regressions, so
it is unlikely that test runs will identify major issues which surfaced in production.

When backported it MUST be accompanied by:

CDPD-27264. HADOOP-17771. S3AFS creation fails "Unable to find a region via the region provider chain."

This patch depends on CDPD-21205, which breaks the HBoss build/test.
Read the instructions there when cherrypicking.

Change-Id: If9b1b1d8ea3556ed4dfb2d164a8a6fa477dfc36d