CDPD-30907. HADOOP-17981. resilient-commit on abfs through etags
This cherrypick pr adds
* EtagSource and contract,
* abfs changes and ability to raise exceptions on rename
But does not contain the changes to FileOutputCommitter.

It has been cherrypicked to reduce merge problems with
the manifest committer and other changes to abfs rename.

-----

Uses a new EtagSource API in hadoop-common (and implemented
supported in ABFS) to implement recovery from failures
during job commit of the form
"java.io.IOException: Failed to rename VersionedFileStatus"

After a rename() call fails, the etag of the file at the
destination path is compared with that of the source file
as determined when listing the parent directory.
If the two match and the source file is no longer present,
the rename is considered successful.

This algorithm relies on
* the etag of files being different on different uploads
  (at a minimum: when different data is uploaded to the same
	destination)
* etag values persisted across renames.
  (true for ADLS Gen2, not for S3)
* directory list API calls returning the etags of all files.

To enable users MUST set
mapreduce.fileoutputcommitter.algorithm.version.v1.experimental.parallel.rename.recovery
to "true"

users SHOULD set
fs.azure.rename.raises.exceptions to true

That is not mandatory, but it gives better error reporting on all forms of rename
failures, and when handling recoverable errors in job commits, reduces the
number of IO requests by one, so reducing load in an IO-intensive phase
of the job.

HADOOP-17981. review javadocs; reduce noisiness

Conflicts:
 hadoop-mapreduce-project/hadoop-mapreduce-client/hadoop-mapreduce-client-core/src/main/java/org/apache/hadoop/mapreduce/lib/output/FileOutputCommitter.java
 hadoop-mapreduce-project/hadoop-mapreduce-client/hadoop-mapreduce-client-core/src/test/java/org/apache/hadoop/mapreduce/lib/output/TestFileOutputCommitter.java

Change-Id: I2bc9aefa0755ed68115526d5f6f3610d9eec6704

Change-Id: I36eb852305634f33a5cfc39b81ac4a5128c14995